refactor(security): flatten security/__tests__ — no basename collisions, clean move (#2565 slice 3)
4 files in security/__tests__ — none had a basename collision at the
top level, so the move is a simple rename + import fix:

- input-validation.test.ts
- path-traversal.test.ts
- rate-limiting.test.ts
- secrets-sanitization.test.ts

Fix:
- `../X` static imports → `./X` (already covered by sed)
- `await import('../../X')` dynamic imports → `await import('../X')`
  (path-traversal.test.ts uses these for vi.mock-resolution patterns)

Remove the now-empty __tests__/ directory.

All 1413 security tests still pass.
